MaghribIA_voiceBot/requirements.txt
livekit-agents>=0.12.1
livekit-plugins-openai>=0.10.9
livekit-plugins-deepgram>=0.6.13
livekit-plugins-silero>=0.7.4
python-dotenv~=1.0
openai==0.27.0
langchain==0.16.0
langchain-openai==0.16.0
langchain-community==0.16.0
livekit==0.3.3
pandas==1.5.3
logging==0.5.1.2
silero==0.0.3
pickle5==0.0.11
Chroma==0.1.6
